finished implementing the LISTGROUP command

+++ src/database/queries/get_group.sql
@@ -1,8 +1,8 @@
 select
   q1.f_group_id,
-  coalesce(q2.f_article_count, 0)::bigint as f_article_count,
-  coalesce(q2.f_low_water_mark, 0)::bigint as f_low_water_mark,
-  coalesce(q2.f_high_water_mark, 0)::bigint as f_high_water_mark
+  coalesce(q2.f_article_count, 0)::integer as f_article_count,
+  coalesce(q2.f_low_water_mark, 0)::integer as f_low_water_mark,
+  coalesce(q2.f_high_water_mark, 0)::integer as f_high_water_mark
 from
   dialogue.t_group q1
   left outer join
